backend/app/pyproject.toml
[tool.poetry]
name = "async_fastapi_bolierplate"
version = "0.1.0"
description = ""
authors = ["Xuan Hien <hiennguyen9874@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.15.0"
python-multipart = "^0.0.5"
email-validator = "^1.1.3"
requests = "^2.26.0"
celery = "^5.1.2"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
tenacity = "^8.0.1"
pydantic = "^1.8.2"
gunicorn = "^20.1.0"
alembic = "^1.7.4"
sentry-sdk = "^1.4.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
asyncpg = "^0.24.0"
psycopg2-binary = "^2.9.1"
SQLAlchemy = "^1.4.26"
importlib-resources = "^5.4.0"
importlib-metadata = "^4.8.1"
redis = "^3.5.3"
fastapi = "^0.68.1"
fastapi-mail = "^1.0.1"
Authlib = "^1.1.0"
starlette = "0.14.2"
itsdangerous = "^2.1.2"

[tool.poetry.dev-dependencies]
black = "^21.9b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
